requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
chromadb
langchain
langchain-google-genai
pydantic-settings
pypdf
streamlit
requests
tenacity
